# src/user-api-base/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "user-api",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:3000"
  } ],
  "paths" : {
    "/users/{userId}" : {
      "get" : {
        "description" : "Retrieve the information of the user with the matching user ID.",
        "operationId" : "get-users-userId",
        "parameters" : [ {
          "description" : "Id of an existing user.",
          "explode" : false,
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : { },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "Get User Alice Smith" : {
                    "value" : {
                      "id" : 142,
                      "firstName" : "Alice",
                      "lastName" : "Smith",
                      "email" : "alice.smith@gmail.com",
                      "dateOfBirth" : "1997-10-31",
                      "emailVerified" : true,
                      "signUpDate" : "2019-08-24"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "User Found"
          },
          "404" : {
            "description" : "User Not Found"
          }
        },
        "summary" : "Get User Info by User ID",
        "tags" : [ ],
        "x-stoplight" : {
          "id" : "xn9js0porvkxo"
        }
      },
      "parameters" : [ {
        "description" : "Id of an existing user.",
        "explode" : false,
        "in" : "path",
        "name" : "userId",
        "required" : true,
        "schema" : { },
        "style" : "simple"
      } ],
      "patch" : {
        "description" : "Update the information of an existing user.",
        "operationId" : "patch-users-userId",
        "parameters" : [ {
          "description" : "Id of an existing user.",
          "explode" : false,
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : { },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "Update First Name" : {
                  "value" : {
                    "firstName" : "Rebecca"
                  }
                },
                "Update Email" : {
                  "value" : {
                    "email" : "rebecca@gmail.com"
                  }
                },
                "Update Last Name & Date of Birth" : {
                  "value" : {
                    "lastName" : "Baker",
                    "dateOfBirth" : "1985-10-02"
                  }
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/patch_users_userId_request"
              }
            }
          },
          "description" : "Patch user properties to update."
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "Updated User Rebecca Baker" : {
                    "value" : {
                      "id" : 13,
                      "firstName" : "Rebecca",
                      "lastName" : "Baker",
                      "email" : "rebecca@gmail.com",
                      "dateOfBirth" : "1985-10-02",
                      "emailVerified" : false,
                      "createDate" : "2019-08-24"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "User Updated"
          },
          "404" : {
            "description" : "User Not Found"
          },
          "409" : {
            "description" : "Email Already Taken"
          }
        },
        "summary" : "Update User Information",
        "x-stoplight" : {
          "id" : "551zsj3hesdn1"
        }
      }
    },
    "/user" : {
      "post" : {
        "description" : "Create a new user.",
        "operationId" : "post-user",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "Create User Bob Fellow" : {
                  "value" : {
                    "firstName" : "Bob",
                    "lastName" : "Fellow",
                    "email" : "bob.fellow@gmail.com",
                    "dateOfBirth" : "1996-08-24"
                  }
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/post_user_request"
              }
            }
          },
          "description" : "Post the necessary fields for the API to create a new user."
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "New User Bob Fellow" : {
                    "value" : {
                      "id" : 12,
                      "firstName" : "Bob",
                      "lastName" : "Fellow",
                      "email" : "bob.fellow@gmail.com",
                      "dateOfBirth" : "1996-08-24",
                      "emailVerified" : false,
                      "createDate" : "2020-11-18"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "User Created"
          },
          "400" : {
            "description" : "Missing Required Information"
          },
          "409" : {
            "description" : "Email Already Taken"
          }
        },
        "summary" : "Create New User",
        "x-stoplight" : {
          "id" : "u7802yhgpoyoi"
        }
      }
    },
    "/users/{userId}/exists" : {
      "get" : {
        "operationId" : "get-users-userId-exists",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : { },
          "style" : "simple"
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "summary" : "Your GET endpoint",
        "tags" : [ ],
        "x-stoplight" : {
          "id" : "ngevk2s29qyhf"
        }
      },
      "parameters" : [ {
        "explode" : false,
        "in" : "path",
        "name" : "userId",
        "required" : true,
        "schema" : { },
        "style" : "simple"
      } ]
    }
  },
  "components" : {
    "schemas" : {
      "User" : {
        "description" : "",
        "example" : {
          "firstName" : "",
          "lastName" : "",
          "emailVerified" : "",
          "dateOfBirth" : "1997-10-31",
          "id" : "",
          "email" : "",
          "createDate" : ""
        },
        "properties" : {
          "id" : {
            "description" : "Unique identifier for the given user.",
            "x-stoplight" : {
              "id" : "q6invr6eh7xi7"
            }
          },
          "firstName" : {
            "x-stoplight" : {
              "id" : "66rat8vxkjdcz"
            }
          },
          "lastName" : {
            "x-stoplight" : {
              "id" : "hf0faoig04o7h"
            }
          },
          "email" : {
            "format" : "email",
            "x-stoplight" : {
              "id" : "wo7mgwjtn1zqm"
            }
          },
          "dateOfBirth" : {
            "example" : "1997-10-31",
            "format" : "date",
            "x-stoplight" : {
              "id" : "35kly88cec8nk"
            }
          },
          "emailVerified" : {
            "description" : "Set to true if the user's email has been verified.",
            "x-stoplight" : {
              "id" : "pyg7u07pumxy0"
            }
          },
          "createDate" : {
            "description" : "The date that the user was created.",
            "format" : "date",
            "x-stoplight" : {
              "id" : "v6hwbxjod5n5h"
            }
          }
        },
        "required" : [ "email", "emailVerified", "firstName", "id", "lastName" ],
        "title" : "User",
        "x-stoplight" : {
          "id" : "joc58ip4trqa5"
        }
      },
      "patch_users_userId_request" : {
        "properties" : {
          "firstName" : { },
          "lastName" : { },
          "email" : {
            "description" : "If a new email is given, the user's email verified property will be set to false."
          },
          "dateOfBirth" : { }
        }
      },
      "post_user_request" : {
        "properties" : {
          "firstName" : { },
          "lastName" : { },
          "email" : { },
          "dateOfBirth" : {
            "format" : "date"
          }
        },
        "required" : [ "dateOfBirth", "email", "firstName", "lastName" ]
      }
    }
  },
  "x-stoplight" : {
    "id" : "807clewcjwj8s"
  }
}